bugfix: allow to filter FloatFields by lt, gt, and range
this requires a new format for FloatFields: the hex-encoded result of xapian.sortable_serialise(). This is necessary because sortable_serialise() returns bytes in Python 3 which we can't pass to xapian.Query. Note that hex-encoded byte strings have the same sort order as the actual byte strings.

Also, use xapian's OP_VALUE_LE and OP_VALUE_GE in the filter_* methods because calling the XHValueRangeProcessor seems too hacky.