xapian_wheel_builder.sh
This adds a script to build a wheel containing xapian python3 bindings
along with a matching build of the xapian library.

I've been using this with xapian v1.4.9 for years now, and recently
merged the linux and macOS versions into a single script.

This should not be used to create wheels for uploading to PyPI, but
works well for building a deployable artefact for a known
target.